Fix simulator with Keras fit_generator (#27)
The batch size is not passed to Keras fit_generator,
but can be passed to Keras fit. When fit_generator
is used the batch_size comes into the Callback's
set_params method with the None value. This None
should not overwrite a batch_size set previously on the
LMS object. This will allow users of fit_generator
to pre-set the batch_size on the LMS object.