Keras ResNet50 - Divide steps by GPU count (#72)
The Keras ResNet50 example should divide the
number of steps per epoch by the number of GPUs
in the distribution module. With non-syntheic data
this would happen by the data set being partitioned
so each GPU operates on a separate portion, and
thus reducing the number of steps in the entire epoch
by a factor of the GPUs.