Allow serialization by topological level memory size (#82)
Allow the selection of topological levels to be serialized by the amount of calculated amount of memory the level will consume.

Current tuning of the serialization parameter is very time consuming and users typically start serializing at the end and binary search to find a value that works. This binary search can take a long time to run and the end result is likely serializing levels that do not need to be serialized for memory usage reasons which can lead to lower data throughput.

These changes allow the selection of topological levels to serialized based on the amount of memory they are calculated to consume. Enabling debug logging at level 1 will output the predicted memory usage of each level.

Tuning serialization with this new parameter will consist of continually decreasing the GiB memory value of the parameter until the training avoids OOM and becomes stable. This should lead to faster wall clock time for tuning as well as faster data throughput.