Compensate for TensorFlow 1.14 changes
TensorFlow 1.14 changes when the optimizer operations are added
to models in the Keras fit_generator flow. It also changes
the named scope of the optimizer operations in Keras models.

For all models, changes to the dependency optimizer cause it
to remove TensorFlow Large Model Support identity swapping nodes.

This compensates for those changes.