Update the header of Keras mnist_cnn based files (#100)
Update the header of the files based on the Keras mnist_cnn example.
This change adds a link to the original file to give some indication
of authorship. It also removes the IBM copyright from the file
so it is not the only copyright present, when in fact, the file is under
other copyrights in its original repository.